Correção das rotas
