+++ modulo4/labex/src/Container/App.jsx
@@ -2,7 +2,7 @@ import { ThemeProvider } from 'styled-components'
 import { GlobalStyles } from '../styles/globalStyles'
 import { theme } from '../styles/themes'
 
-import { Router } from '../routes/routes'
+import { Router } from '../routes/Routes'
 
 const App = () => {
   return (
