+++ modulo4/labex/src/pages/Admin/Login/index.jsx
@@ -1,13 +1,13 @@
 import * as Styled from './styles'
 import { useNavigate } from 'react-router-dom'
-import { goAdminTripDetails, goBack } from '../../../routes/coordinator'
+import { goAdminHome, goBack } from '../../../routes/coordinator'
 
 export const Login = () => {
   const navigate = useNavigate()
 
   return (
     <>
-      <button onClick={() => goAdminTripDetails(navigate)}>Ir para home admin</button>
+      <button onClick={() => goAdminHome(navigate)}>Ir para home admin</button>
       <button onClick={() => goBack(navigate)}>Voltar página anterior</button>
     </>
   )
