+++ modulo4/labex/src/pages/User/ApplicationForm/index.jsx
@@ -1,6 +1,6 @@
 import * as Styled from './styles'
 import { useNavigate } from 'react-router-dom'
-import { goBack } from '../../routes/coordinator'
+import { goBack } from '../../../routes/coordinator'
 
 export const ApplicationForm = () => {
   const navigate = useNavigate()
